Move macro handling tests into separate file
